Quick Game set picker: reuse the tournament lobby's searchable modal
The Quick Game lobby's "Random" deck tab used a bare native <select> for
choosing a sealed-pool set — no search, icons, partial-set filtering, or
block grouping. The tournament lobby already has a far richer set browser
inline in GameUI.

Extract that browser into a shared `SetPickerModal` component and use it in
both places:

- New `SetPickerModal.tsx` owns search / show-partial-sets toggle / min-cards
  filter / block-grouped, newest-first rows with set icons, release dates and
  card counts. Supports `multi` (toggle, stays open — tournament) and `single`
  (select-and-close, with an optional "Random Set" reset row — Quick Game).
- `GameUI.tsx` drops its inline picker state/helpers/JSX and renders
  `<SetPickerModal mode="multi">`. No visual change to the tournament lobby.
- `DeckPicker.tsx` Random tab replaces the <select> with a set chip + "Choose
  set" button that opens `<SetPickerModal mode="single" allowRandom>`, matching
  the lobby's chip styling. `availableSets` widened to the full `AvailableSet`
  (the Quick Game store already supplies the rich metadata).
